Fix bug in single-commit branch with update commits
If a branch contains a single non-conventional commit plus merge commits from its base branch (i.e. if the update button in the PR is used), GitHub ignores these merge commits for the purpose of producing the squashed commit's message.

This means that GitHub will merge a multi-commit branch with the message of that single non-conventional commit.

This change updates the rules such that merge-commits-from-master are ignored for the purpose of determining if a branch contains a single commit.